Remove subject from group (#287)
* adjusting subject delete to only remove a subject from a protocol, not delete the subject. It might be the case that this subject is on another protocol.

* making delete subject api call for super users only

+++ api/views/protocol.py
@@ -452,15 +452,35 @@ class ProtocolSubjectDetailView(BRPApiView):
         )
 
     def delete(self, request, pk, subject, *args, **kwargs):
-        try:
-            subject = self.s_rh.get(id=subject)
-            protocol = Protocol.objects.get(pk=pk)
-            self.s_rh.delete(id=subject.id)
-            SubjectUtils.delete_protocol_subject_record_group(protocol, subject)
-        except:
-            return Response({'error': 'Unable to delete subject'}, status=400)
+        if request.user.is_superuser:
+            try:
+                subject = self.s_rh.get(id=subject)
+                protocol = Protocol.objects.get(pk=pk)
+                group_id = self.get_group_id(protocol.immutable_key.key)
+                # instead of deleting a subject - we want to just remove them from the subject group
+                # self.s_rh.delete(id=subject.id)
+                self.remove_sub_from_ehb_group(subject.id, group_id)
+                SubjectUtils.delete_protocol_subject_record_group(protocol, subject)
+            except:
+                return Response({'error': 'Unable to delete subject'}, status=400)
 
-        return Response({'info': 'Subject deleted'}, status=200)
+            return Response({'info': 'Subject deleted'}, status=200)
+        else:
+            return Response(
+                {"detail": "You are not authorized to view subjects in this protocol"},
+                status=403
+            )
+
+    @staticmethod
+    def get_group_id(key):
+        get_group_url = '/api/group/?name=' +'BRP:' + key
+        group = ServiceClient.ehb_api(get_group_url, "GET").json()
+        return group['id']
+
+    @staticmethod
+    def remove_sub_from_ehb_group(sub_pk, group_pk):
+        url = '/api/group/id/' + str(group_pk) + '/subjects/id/' + str(sub_pk) + '/'
+        return ServiceClient.ehb_api(url, "DELETE")
 
     @staticmethod
     def updateEhbSubject(subjectID, new_subject, old_subject):
